backend/requirements.txt
asgiref==3.9.1
certifi==2025.7.14
cloudinary==1.44.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
